# Speed up hamming distance for bit arrays