@timholy 2024-07-04
Another way it's potentially slow is because it allocates a temporary. Have you evaluated
```julia
sum((a, b) -> xor(a, b), zip(desc_1, desc_2)) / length(desc_1)
```
Another option would be to write out the loop; the above relies on a generator and we don't fully specialize generators (at least not yet). Writing it out as a loop is likely to be your fastest option, if you are really performance sensitive.