@ymtoo 2024-07-05
```
hamming_distance(desc_1, desc_2) = sum((a, b) -> xor(a, b), zip(desc_1, desc_2)) / length(desc_1)
```
throws an error
```
ERROR: MethodError: no method matching (::ImageFeatures.var"#115#116")(::Tuple{Bool, Bool})

Closest candidates are:
  (::ImageFeatures.var"#115#116")(::Any, ::Any)
   @ ImageFeatures ~/Projects/ImageFeatures.jl/src/core.jl:78

Stacktrace:
  [1] MappingRF
    @ ./reduce.jl:100 [inlined]
  [2] _foldl_impl
    @ ./reduce.jl:58 [inlined]
  [3] foldl_impl(op::Base.MappingRF{…}, nt::Base._InitialValue, itr::Base.Iterators.Zip{…})
    @ Base ./reduce.jl:48
  [4] mapfoldl_impl(f::ImageFeatures.var"#115#116", op::typeof(Base.add_sum), nt::Base._InitialValue, itr::Base.Iterators.Zip{…})
    @ Base ./reduce.jl:44
  [5] mapfoldl(f::Function, op::Function, itr::Base.Iterators.Zip{Tuple{BitVector, BitVector}}; init::Base._InitialValue)
    @ Base ./reduce.jl:175
  [6] mapfoldl
    @ ./reduce.jl:175 [inlined]
  [7] mapreduce
    @ ./reduce.jl:307 [inlined]
  [8] sum(f::Function, a::Base.Iterators.Zip{Tuple{BitVector, BitVector}})
    @ Base ./reduce.jl:535
  [9] hamming_distance(desc_1::BitVector, desc_2::BitVector)
    @ ImageFeatures ~/Projects/ImageFeatures.jl/src/core.jl:78
 [10] top-level scope
    @ REPL[25]:1
Some type information was truncated. Use `show(err)` to see complete types.
```
This
```
hamming_distance(desc_1, desc_2) = sum(ab -> xor(ab...), zip(desc_1, desc_2)) / length(desc_1)
```
works but it is slow as compared to `sum(x)/length(x)`.
```
julia> @btime hamming_distance($b1, $b2)
  677.013 ns (0 allocations: 0 bytes)
0.54296875
```