@timholy 2024-07-08
Thanks for checking. There may be an even faster option, but let's go with this for now.

Thanks for the contribution!